% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_gene_score.R
\name{aggregate_gene_score}
\alias{aggregate_gene_score}
\title{Compute an aggregate gene signature score}
\usage{
aggregate_gene_score(cds, genes, col_name = "gene_signature_score")
}
\arguments{
\item{cds}{monocle3 cell_data_set}

\item{genes}{Character vector of gene symbols}

\item{col_name}{Name of output column}
}
\value{
CDS with aggregate score added to colData
}
\description{
Computes a scaled aggregate score for a gene set using
size-factor–normalized expression.
}
