% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_species_conversion.R
\name{convert_human_to_mouse}
\alias{convert_human_to_mouse}
\title{Convert a human gene list to mouse orthologs}
\usage{
convert_human_to_mouse(genes)
}
\arguments{
\item{genes}{Character vector of human gene symbols (HGNC)}
}
\value{
Character vector of mouse gene symbols (MGI)
}
\description{
Uses Ensembl BioMart to map human HGNC symbols to mouse MGI symbols.
}
