% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_species_conversion.R
\name{convert_mouse_to_human}
\alias{convert_mouse_to_human}
\title{Convert a mouse gene list to human orthologs}
\usage{
convert_mouse_to_human(genes)
}
\arguments{
\item{genes}{Character vector of mouse gene symbols (MGI)}
}
\value{
Character vector of human gene symbols (HGNC)
}
\description{
Uses Ensembl BioMart to map mouse MGI symbols to human HGNC symbols.
}
