% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_GEO.R
\name{find_geo_samples}
\alias{find_geo_samples}
\title{Discover GEO matrix samples in a directory}
\usage{
find_geo_samples(folder_path)
}
\arguments{
\item{folder_path}{Path containing *_matrix.mtx.gz files}
}
\value{
Character vector of sample prefixes
}
\description{
Discover GEO matrix samples in a directory
}
