% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_matrix.R
\name{get_expr_matrix}
\alias{get_expr_matrix}
\title{Retrieve an expression matrix}
\usage{
get_expr_matrix(cds, assay = "logcounts")
}
\arguments{
\item{cds}{A \code{SingleCellExperiment} or Monocle3
\code{cell_data_set} object.}

\item{assay}{Name of the assay to retrieve.}
}
\value{
A matrix-like object containing expression values.
}
\description{
Extract an expression matrix from a
\code{SingleCellExperiment}-derived object (including Monocle3
\code{cell_data_set} objects). If the requested assay is
`"logcounts"` but no log-normalized assay is present, a
log-transformed count matrix is generated on the fly using
`log1p(counts(cds))`.
}
\examples{
expr <- get_expr_matrix(cds)
expr <- get_expr_matrix(cds, assay = "counts")

}
