% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_GEO.R
\name{load_geo_sample}
\alias{load_geo_sample}
\title{Load a single GEO matrix into a monocle3 CDS}
\usage{
load_geo_sample(sample_name, folder_path)
}
\arguments{
\item{sample_name}{Sample prefix (no suffix)}

\item{folder_path}{Path containing GEO files}
}
\value{
monocle3 cell_data_set
}
\description{
Load a single GEO matrix into a monocle3 CDS
}
