% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ma_expression.R
\name{plot_ma_expression}
\alias{plot_ma_expression}
\title{Plot an MA expression plot}
\usage{
plot_ma_expression(
  expr1,
  expr2,
  shared_genes,
  top_n = 50,
  labels = c("Dataset 1", "Dataset 2")
)
}
\arguments{
\item{expr1}{A matrix-like object containing expression values.}

\item{expr2}{A matrix-like object containing expression values.}

\item{shared_genes}{Character vector of genes present in both datasets.}

\item{top_n}{Number of highly expressed genes to label from each dataset.}

\item{labels}{Character vector of length two giving the dataset names.}
}
\value{
A \code{ggplot} object.
}
\description{
Create an MA plot comparing the mean expression of shared genes between two
expression matrices. The most highly expressed genes in each dataset are
highlighted and labelled.
}
\examples{
plot_ma_expression(expr_mouse, expr_human, shared_genes)

}
