% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qc_violins.R
\name{plot_qc_violins}
\alias{plot_qc_violins}
\title{Plot QC metrics as violin plots}
\usage{
plot_qc_violins(
  metadata_df,
  metrics,
  metric_labels = NULL,
  plot_titles = NULL,
  group_col = NULL,
  y_limits = NULL,
  violin_color = "#4C72B0",
  point_size = 0,
  base_size = 18
)
}
\description{
Plot QC metrics as violin plots
}
