% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_bar_from_meta.R
\name{plot_stacked_bar_from_meta}
\alias{plot_stacked_bar_from_meta}
\title{Plot stacked bar chart from metadata}
\usage{
plot_stacked_bar_from_meta(
  meta1,
  meta2,
  col_order = NULL,
  fill_by = "meta1",
  fill_colors = NULL,
  fill_order = NULL,
  fill_title = "legend_title",
  xlab = "cell type",
  ylab = "percentage of cells",
  normalize = c("none", "sample", "celltype", "both")
)
}
\description{
Plot stacked bar chart from metadata
}
