% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_manustyle.R
\name{plot_umap_manustyle}
\alias{plot_umap_manustyle}
\title{Plot UMAP with custom styling}
\usage{
plot_umap_manustyle(
  cds,
  celltype_colors = NULL,
  color_cells_by = "fine_annotations",
  group_cells_by = "fine_annotations",
  palette = scales::hue_pal(),
  cell_size = 0.1,
  label_size = 1.5,
  label_fill_alpha = 0.7,
  show_trajectory_graph = FALSE,
  label_cell_groups = FALSE
)
}
\description{
Plot UMAP with custom styling
}
