% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrublet.R
\name{run_scrublet}
\alias{run_scrublet}
\title{Run Scrublet doublet detection on a CDS}
\usage{
run_scrublet(cds, expected_doublet_rate = 0.06, conda_env = NULL)
}
\arguments{
\item{cds}{monocle3 cell_data_set}

\item{expected_doublet_rate}{Numeric}

\item{conda_env}{Conda environment name}
}
\value{
CDS with scrublet_score and scrublet_call
}
\description{
Run Scrublet doublet detection on a CDS
}
