% Generated by roxygen2: do not edit by hand
\name{subset_xenium_rois}
\alias{subset_xenium_rois}
\title{Subset a Xenium CDS to cells falling inside selected ROIs}
\usage{subset_xenium_rois(cds, rois, sample_col = "sample", x_col = "x_centroid", y_col = "y_centroid", roi_col = "roi_id")}
\arguments{
\item{cds}{A monocle3 cell_data_set.}
\item{rois}{ROI table from \code{select_xenium_rois()}.}
\item{sample_col, x_col, y_col}{Column names in \code{colData(cds)}.}
\item{roi_col}{Name of the added ROI membership column.}
}
\value{A subsetted cell_data_set with ROI membership in \code{colData}.}
\description{Subset a Xenium CDS to cells falling inside selected ROIs.}
